requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
